+++ Runtime/Scripts/Utils/IPRuntimeEditorCore.cs
@@ -1,0 +1,15 @@
+/*
+ *	Created by:  Peter @sHTiF Stefcek
+ */
+
+#if UNITY_EDITOR
+namespace InstancePainter
+{
+    public class IPRuntimeEditorCore
+    {
+        public static bool renderingAsUtil = false;
+        
+        public static ICluster explicitCluster;
+    }
+}
+#endif